Add the ability to use a ledger device with the mint command (#5417)
